pyproject.toml
[project]
name = "lytro-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "opencv-python>=4.13.0.92",
    "pyusb>=1.3.1",
    "typed-argument-parser>=1.11.0",
]
